Struct glium_pib::WindowConfig

pub struct WindowConfig {
    pub display: Display,
    pub surface_size: Option<(u32, u32)>,
    pub red: ColorBits,
    pub green: ColorBits,
    pub blue: ColorBits,
    pub alpha: Option<ColorBits>,
    pub depth: Option<DepthBits>,
}

Window configuration.

Fields

display

The display to use.

surface_size

The size of the surface to render to. If none, the size of the display is used.

red

Number of bits per pixel used for the red channel.

green

Number of bits per pixel used for the green channel.

blue

Number of bits per pixel used for the blue channel.

alpha

Number of bits per pixel used for the alpha channel.

depth

Number of bits per pixel used for the depth buffer.